+++ filenames.gni
@@ -482,6 +482,7 @@ filenames = {
     "shell/common/gin_converters/image_converter_gin_adapter.h",
     "shell/common/gin_converters/message_box_converter.cc",
     "shell/common/gin_converters/message_box_converter.h",
+    "shell/common/gin_converters/native_window_converter.h",
     "shell/common/gin_converters/net_converter_gin_adapter.h",
     "shell/common/gin_converters/string16_converter.h",
     "shell/common/gin_util.h",
@@ -513,6 +514,7 @@ filenames = {
     "shell/common/native_mate_converters/image_converter.cc",
     "shell/common/native_mate_converters/image_converter.h",
     "shell/common/native_mate_converters/map_converter.h",
+    "shell/common/native_mate_converters/native_window_converter.h",
     "shell/common/native_mate_converters/net_converter.cc",
     "shell/common/native_mate_converters/net_converter.h",
     "shell/common/native_mate_converters/network_converter.cc",
