+++ shell/browser/api/atom_api_top_level_window.h
@@ -261,43 +261,4 @@ class TopLevelWindow : public mate::TrackableObject<TopLevelWindow>,
 
 }  // namespace electron
 
-namespace gin {
-
-template <>
-struct Converter<electron::NativeWindow*> {
-  static bool FromV8(v8::Isolate* isolate,
-                     v8::Local<v8::Value> val,
-                     electron::NativeWindow** out) {
-    // null would be tranfered to NULL.
-    if (val->IsNull()) {
-      *out = NULL;
-      return true;
-    }
-
-    electron::api::TopLevelWindow* window;
-    // TODO(deermichel): remove mate:: after dropping mate
-    if (!mate::Converter<electron::api::TopLevelWindow*>::FromV8(isolate, val,
-                                                                 &window))
-      return false;
-    *out = window->window();
-    return true;
-  }
-};
-
-}  // namespace gin
-
-// TODO(deermichel): remove after dropping mate
-namespace mate {
-
-template <>
-struct Converter<electron::NativeWindow*> {
-  static bool FromV8(v8::Isolate* isolate,
-                     v8::Local<v8::Value> val,
-                     electron::NativeWindow** out) {
-    return gin::ConvertFromV8(isolate, val, out);
-  }
-};
-
-}  // namespace mate
-
 #endif  // SHELL_BROWSER_API_ATOM_API_TOP_LEVEL_WINDOW_H_
