+++ shell/common/gin_converters/message_box_converter.cc
@@ -5,8 +5,8 @@
 #include "shell/common/gin_converters/message_box_converter.h"
 
 #include "gin/dictionary.h"
-#include "shell/browser/api/atom_api_top_level_window.h"
 #include "shell/common/gin_converters/image_converter_gin_adapter.h"
+#include "shell/common/gin_converters/native_window_converter.h"
 
 namespace gin {
